+++ tests/InjectBaseAbstract.php
@@ -2,9 +2,10 @@
 
 namespace Mizmoz\Container\Tests;
 
+use Mizmoz\Container\Contract\ManageContainerInterface;
 use Mizmoz\Container\ManageContainerTrait;
 
-abstract class InjectBaseAbstract
+abstract class InjectBaseAbstract implements ManageContainerInterface
 {
     use ManageContainerTrait;
 }