+++ tests/ResolveLoggerInterface.php
@@ -4,5 +4,11 @@ namespace Mizmoz\Container\Tests;
 
 interface ResolveLoggerInterface
 {
-    public function log(string $message);
+    /**
+     * Log a message
+     *
+     * @param string $message
+     * @return void
+     */
+    public function log(string $message): void;
 }