+++ tests/ResolveNoConstructor.php
@@ -2,9 +2,10 @@
 
 namespace Mizmoz\Container\Tests;
 
+use Mizmoz\Container\Contract\ManageContainerInterface;
 use Mizmoz\Container\ManageContainerTrait;
 
-class ResolveNoConstructor
+class ResolveNoConstructor implements ManageContainerInterface
 {
     use ManageContainerTrait;
 }