+++ tests/TestCase.php
@@ -4,7 +4,8 @@ use Mockery;
 
 class TestCase extends \PHPUnit\Framework\TestCase
 {
-    public function tearDown() {
+    public function tearDown(): void
+    {
         if ($container = Mockery::getContainer()) {
             $this->addToAssertionCount($container->mockery_getExpectationCount());
         }